code cleanup

+++ src/main/java/com/esaulpaugh/headlong/util/FastHex.java
@@ -35,7 +35,7 @@ public final class FastHex {
     private static final byte NO_MAPPING = -1;
 
     static {
-        final short[] ints = new short[] { '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f' };
+        final byte[] ints = new byte[] { '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f' };
         final int leftNibbleMask = 0xF0;
         final int rightNibbleMask = 0x0F;
         for (int i = 0; i < ENCODE_TABLE.length; i++) {
